requirements.txt
alembic==1.0.7
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autobahn==19.1.1
autoenv==1.0.0
Automat==0.7.0
backcall==0.1.0
buildbot==1.8.0
buildbot-worker==1.8.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
constantly==15.1.0
coverage==4.5.2
cryptography==2.5
decorator==4.3.2
et-xmlfile==1.0.1
Flask==1.0.2
flask-marshmallow==0.9.0
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
future==0.17.1
gunicorn==19.9.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10
jupyter-client==5.2.4
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.18.0
marshmallow-sqlalchemy==0.16.0
mccabe==0.6.1
mock==2.0.0
more-itertools==5.0.0
numpy==1.16.3
openpyxl==2.6.2
pandas==0.24.2
parso==0.4.0
pathlib2==2.3.3
pbr==5.1.2
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.8.1
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.2.2
pyOpenSSL==19.0.0
pytest==4.1.1
pytest-cov==2.6.1
python-dateutil==2.7.5
python-editor==1.0.4
pytz==2019.1
PyYAML==3.13
pyzmq==18.0.1
requests==2.21.0
rope==0.12.0
service-identity==18.1.0
six==1.12.0
SQLAlchemy==1.2.17
sqlalchemy-migrate==0.12.0
sqlparse==0.2.4
stringcase==1.2.0
Tempita==0.5.2
tornado==6.0.2
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.2.0
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.1
xlrd==1.2.0
XlsxWriter==1.1.7
zope.interface==4.6.0
